+++ app/static/tests/unit/components/run/runTab.test.ts
@@ -5,8 +5,9 @@ jest.mock("plotly.js", () => {});
 import Vuex from "vuex";
 import { shallowMount } from "@vue/test-utils";
 import { BasicState } from "../../../../src/app/store/basic/state";
-import { mockBasicState, mockModelState } from "../../../mocks";
+import { mockBasicState, mockModelState, mockRunState } from "../../../mocks";
 import { ModelState } from "../../../../src/app/store/model/state";
+import { RunState } from "../../../../src/app/store/run/state";
 import RunTab from "../../../../src/app/components/run/RunTab.vue";
 import RunPlot from "../../../../src/app/components/run/RunPlot.vue";
 import ErrorInfo from "../../../../src/app/components/ErrorInfo.vue";
@@ -16,8 +17,11 @@ describe("RunTab", () => {
     const defaultModelState = {
         odinRunner: {} as any,
         odin: {} as any,
-        runRequired: false,
         compileRequired: false
+    };
+
+    const defaultRunState = {
+        runRequired: false
     };
 
     beforeEach(() => {
@@ -26,13 +30,20 @@ describe("RunTab", () => {
 
     const mockRunModel = jest.fn();
 
-    const getWrapper = (modelState: Partial<ModelState> = defaultModelState) => {
+    const getWrapper = (modelState: Partial<ModelState> = defaultModelState,
+                        runState: Partial<RunState> = defaultRunState) => {
         const store = new Vuex.Store<BasicState>({
             state: mockBasicState(),
             modules: {
                 model: {
                     namespaced: true,
                     state: mockModelState(modelState),
+                    actions: {
+                    }
+                },
+                run: {
+                    namespaced: true,
+                    state: mockRunState(runState),
                     actions: {
                         RunModel: mockRunModel
                     }
@@ -66,7 +77,6 @@ describe("RunTab", () => {
 
     it("disabled run button when compile is required", () => {
         const wrapper = getWrapper({
-            runRequired: false,
             compileRequired: true
         });
         expect(wrapper.find("button").element.disabled).toBe(true);
@@ -74,7 +84,6 @@ describe("RunTab", () => {
 
     it("fades plot and shows message when compile required", () => {
         const wrapper = getWrapper({
-            runRequired: false,
             compileRequired: true
         });
         expect(wrapper.findComponent(ActionRequiredMessage).props("message")).toBe(
@@ -84,10 +93,7 @@ describe("RunTab", () => {
     });
 
     it("fades plot and shows message when model run required", () => {
-        const wrapper = getWrapper({
-            runRequired: true,
-            compileRequired: false
-        });
+        const wrapper = getWrapper({ compileRequired: false }, { runRequired: true } );
         expect(wrapper.findComponent(ActionRequiredMessage).props("message")).toBe(
             "Model code has been recompiled or options have been updated. Run Model to view updated graph."
         );
@@ -102,7 +108,7 @@ describe("RunTab", () => {
 
     it("displays error info in run model", () => {
         const odinRunnerError = { error: "model error", detail: "with details" };
-        const wrapper = getWrapper({ odinRunnerError });
+        const wrapper = getWrapper({}, { error: odinRunnerError });
         expect(wrapper.findComponent(ErrorInfo).exists()).toBe(true);
         expect(wrapper.findComponent(ErrorInfo).props("error")).toStrictEqual(odinRunnerError);
     });
