+++ app/static/tests/unit/store/model/actions.test.ts
@@ -1,13 +1,15 @@
 import Vuex from "vuex";
 import {
-    mockAxios, mockBasicState, mockCodeState, mockFailure, mockModelState, mockSuccess
+    mockAxios, mockBasicState, mockCodeState, mockFailure, mockModelState, mockRunState, mockRunner, mockSuccess
 } from "../../../mocks";
 import { actions, ModelAction } from "../../../../src/app/store/model/actions";
 import { ModelMutation, mutations } from "../../../../src/app/store/model/mutations";
 import { BasicState } from "../../../../src/app/store/basic/state";
 import { AppType } from "../../../../src/app/store/appState/state";
+import { RunAction, actions as runActions } from "../../../../src/app/store/run/actions";
 import { FitDataAction } from "../../../../src/app/store/fitData/actions";
 import { ModelFitAction } from "../../../../src/app/store/modelFit/actions";
+import { RunMutation, mutations as runMutations } from "../../../../src/app/store/run/mutations";
 import { ModelFitMutation } from "../../../../src/app/store/modelFit/mutations";
 import { SensitivityMutation } from "../../../../src/app/store/sensitivity/mutations";
 
@@ -25,13 +27,15 @@ describe("Model actions", () => {
             paramSettings: {
                 parameterToVary: null
             }
+        },
+        run: {
+            runRequired: false,
+            parameterValues: new Map([
+                ["p1", 1],
+                ["p2", 2],
+                ["p3", 3]
+            ])
         }
-    };
-
-    const mockRunner = () => {
-        return {
-            wodinRun: jest.fn((odin, pars, start, end) => "test solution" as any)
-        } as any;
     };
 
     it("fetches odin runner", async () => {
@@ -100,13 +104,7 @@ describe("Model actions", () => {
                     variables: ["x", "y"]
                 }
             },
-            compileRequired: true,
-            runRequired: false,
-            parameterValues: new Map([
-                ["p1", 1],
-                ["p2", 2],
-                ["p3", 3]
-            ])
+            compileRequired: true
         };
         const commit = jest.fn();
         const dispatch = jest.fn();
@@ -114,7 +112,7 @@ describe("Model actions", () => {
         expect(commit.mock.calls.length).toBe(7);
         expect(commit.mock.calls[0][0]).toBe(ModelMutation.SetOdin);
         expect(commit.mock.calls[0][1]).toBe(3);
-        expect(commit.mock.calls[1][0]).toBe(ModelMutation.SetParameterValues);
+        expect(commit.mock.calls[1][0]).toBe(`run/${RunMutation.SetParameterValues}`);
         const expectedParams = new Map([["p2", 20], ["p3", 30], ["p4", 40]]);
         expect(commit.mock.calls[1][1]).toStrictEqual(expectedParams);
         expect(commit.mock.calls[2][0]).toBe(ModelMutation.SetPaletteModel);
@@ -122,7 +120,7 @@ describe("Model actions", () => {
         // might need a SetCompileRequired false here too?
         expect(commit.mock.calls[3][0]).toBe(ModelMutation.SetCompileRequired);
         expect(commit.mock.calls[3][1]).toBe(false);
-        expect(commit.mock.calls[4][0]).toBe(ModelMutation.SetRunRequired);
+        expect(commit.mock.calls[4][0]).toBe(`run/${RunMutation.SetRunRequired}`);
         expect(commit.mock.calls[4][1]).toBe(true);
         expect(commit.mock.calls[5][0]).toBe(`sensitivity/${SensitivityMutation.SetUpdateRequired}`);
         expect(commit.mock.calls[5][1]).toBe(true);
@@ -144,9 +142,7 @@ describe("Model actions", () => {
                     variables: ["x", "y"]
                 }
             },
-            compileRequired: true,
-            runRequired: false,
-            parameterValues: {}
+            compileRequired: true
         };
         const fitRootState = {
             appType: AppType.Fit,
@@ -163,10 +159,10 @@ describe("Model actions", () => {
         });
         expect(commit.mock.calls.length).toBe(8);
         expect(commit.mock.calls[0][0]).toBe(ModelMutation.SetOdin);
-        expect(commit.mock.calls[1][0]).toBe(ModelMutation.SetParameterValues);
+        expect(commit.mock.calls[1][0]).toBe(`run/${RunMutation.SetParameterValues}`);
         expect(commit.mock.calls[2][0]).toBe(ModelMutation.SetPaletteModel);
         expect(commit.mock.calls[3][0]).toBe(ModelMutation.SetCompileRequired);
-        expect(commit.mock.calls[4][0]).toBe(ModelMutation.SetRunRequired);
+        expect(commit.mock.calls[4][0]).toBe(`run/${RunMutation.SetRunRequired}`);
         expect(commit.mock.calls[5][0]).toBe(`sensitivity/${SensitivityMutation.SetUpdateRequired}`);
         expect(commit.mock.calls[5][1]).toBe(true);
         expect(commit.mock.calls[6][0]).toBe(`sensitivity/${SensitivityMutation.SetParameterToVary}`);
@@ -216,10 +212,10 @@ describe("Model actions", () => {
         });
         expect(commit.mock.calls.length).toBe(6);
         expect(commit.mock.calls[0][0]).toBe(ModelMutation.SetOdin);
-        expect(commit.mock.calls[1][0]).toBe(ModelMutation.SetParameterValues);
+        expect(commit.mock.calls[1][0]).toBe(`run/${RunMutation.SetParameterValues}`);
         expect(commit.mock.calls[2][0]).toBe(ModelMutation.SetPaletteModel);
         expect(commit.mock.calls[3][0]).toBe(ModelMutation.SetCompileRequired);
-        expect(commit.mock.calls[4][0]).toBe(ModelMutation.SetRunRequired);
+        expect(commit.mock.calls[4][0]).toBe(`run/${RunMutation.SetRunRequired}`);
         expect(commit.mock.calls[5][0]).toBe(`sensitivity/${SensitivityMutation.SetUpdateRequired}`);
     });
 
@@ -232,15 +228,14 @@ describe("Model actions", () => {
                     variables: ["x", "y"]
                 }
             } as any,
-            compileRequired: false,
-            runRequired: true
+            compileRequired: false
         });
         const commit = jest.fn();
         (actions[ModelAction.CompileModel] as any)({ commit, state, rootState });
         expect(commit.mock.calls.length).toBe(4);
         expect(commit.mock.calls[0][0]).toBe(ModelMutation.SetOdin);
         expect(commit.mock.calls[0][1]).toBe(3);
-        expect(commit.mock.calls[1][0]).toBe(ModelMutation.SetParameterValues);
+        expect(commit.mock.calls[1][0]).toBe(`run/${RunMutation.SetParameterValues}`);
         expect(commit.mock.calls[1][1]).toStrictEqual(new Map());
         expect(commit.mock.calls[3][0]).toBe(`sensitivity/${SensitivityMutation.SetParameterToVary}`);
         expect(commit.mock.calls[3][1]).toBe(null);
@@ -253,115 +248,8 @@ describe("Model actions", () => {
         expect(commit.mock.calls.length).toBe(0);
     });
 
-    it("runs model and updates required action", () => {
-        const mockOdin = {} as any;
-
-        const parameterValues = new Map([["p1", 1], ["p2", 2]]);
-        const runner = mockRunner();
-        const state = mockModelState({
-            odinRunner: runner,
-            odin: mockOdin,
-            compileRequired: false,
-            runRequired: true,
-            parameterValues,
-            endTime: 99
-        });
-        const commit = jest.fn();
-
-        (actions[ModelAction.RunModel] as any)({ commit, state });
-
-        const run = runner.wodinRun;
-        expect(run.mock.calls[0][0]).toBe(mockOdin);
-        expect(run.mock.calls[0][1]).toStrictEqual(parameterValues);
-        expect(run.mock.calls[0][2]).toBe(0); // start
-        expect(run.mock.calls[0][3]).toBe(99); // end time from state
-
-        expect(commit.mock.calls.length).toBe(2);
-        expect(commit.mock.calls[0][0]).toBe(ModelMutation.SetOdinSolution);
-        expect(commit.mock.calls[0][1]).toBe("test solution");
-        expect(commit.mock.calls[1][0]).toBe(ModelMutation.SetRunRequired);
-        expect(commit.mock.calls[1][1]).toBe(false);
-    });
-
-    it("run model does not update required action if required action was not run", () => {
-        const mockOdin = {} as any;
-
-        const state = mockModelState({
-            odinRunner: mockRunner(),
-            odin: mockOdin,
-            compileRequired: true,
-            runRequired: false,
-            parameterValues: new Map()
-        });
-        const commit = jest.fn();
-
-        (actions[ModelAction.RunModel] as any)({ commit, state });
-        expect(commit.mock.calls.length).toBe(1);
-        expect(commit.mock.calls[0][0]).toBe(ModelMutation.SetOdinSolution);
-    });
-
-    it("run model does nothing if odin runner is not set", () => {
-        const mockOdin = {} as any;
-
-        const state = mockModelState({
-            odinRunner: null,
-            odin: mockOdin
-        });
-        const commit = jest.fn();
-
-        (actions[ModelAction.RunModel] as any)({ commit, state });
-
-        expect(commit).not.toHaveBeenCalled();
-    });
-
-    it("run model does nothing if odin is not set", () => {
-        const runner = mockRunner();
-        const state = mockModelState({
-            odinRunner: runner,
-            odin: null
-        });
-        const commit = jest.fn();
-
-        (actions[ModelAction.RunModel] as any)({ commit, state });
-
-        expect(commit).not.toHaveBeenCalled();
-        expect(runner.wodinRun).not.toHaveBeenCalled();
-    });
-
-    it("runs model throws exception when error in code run model", () => {
-        const mockError = new Error("test");
-        const mockOdin = {} as any;
-        const mockRunnerWithThrownException = () => {
-            return {
-                wodinRun: jest.fn().mockImplementation(() => {
-                    throw mockError;
-                })
-            } as any;
-        };
-
-        const parameterValues = new Map([["p1", 1], ["p2", 2]]);
-        const runner = mockRunnerWithThrownException();
-        const state = mockModelState({
-            odinRunner: runner,
-            odin: mockOdin,
-            compileRequired: false,
-            runRequired: true,
-            parameterValues,
-            endTime: 99
-        });
-        const commit = jest.fn();
-
-        (actions[ModelAction.RunModel] as any)({ commit, state });
-
-        expect(runner.wodinRun.mock.calls[0][0]).toBe(mockOdin);
-        expect(commit.mock.calls.length).toBe(2);
-        expect(commit.mock.calls[0][0]).toBe(ModelMutation.SetOdinRunnerError);
-        expect(commit.mock.calls[0][1]).toStrictEqual({
-            detail: mockError,
-            error: "An error occurred while running the model"
-        });
-    });
-
+    // This one needs some work as we don't kick off the run after
+    // compile, and that causes the mocks to be very confused.
     it("DefaultModel fetches, compiles and runs default model synchronously", async () => {
         // Use real store so can trace the flow of updates through the state
         const runner = mockRunner();
@@ -378,10 +266,17 @@ describe("Model actions", () => {
                     namespaced: true,
                     state: mockModelState({
                         odinRunner: runner,
-                        endTime: 99
                     }),
                     mutations,
                     actions
+                },
+                run: {
+                    namespaced: true,
+                    state: mockRunState({
+                        endTime: 99
+                    }),
+                    mutations: runMutations,
+                    actions: runActions
                 },
                 sensitivity: {
                     namespaced: true,
@@ -405,6 +300,7 @@ describe("Model actions", () => {
         const commit = jest.spyOn(store, "commit");
 
         await store.dispatch(`model/${ModelAction.DefaultModel}`);
+
         expect(commit.mock.calls.length).toBe(11);
 
         // fetch
@@ -419,7 +315,7 @@ describe("Model actions", () => {
         // compile
         expect(commit.mock.calls[2][0]).toBe(`model/${ModelMutation.SetOdin}`);
         expect(commit.mock.calls[2][1]).toBe(3); // evaluated value of test model
-        expect(commit.mock.calls[3][0]).toBe(`model/${ModelMutation.SetParameterValues}`);
+        expect(commit.mock.calls[3][0]).toBe(`run/${RunMutation.SetParameterValues}`);
         expect(commit.mock.calls[3][1]).toStrictEqual(new Map([["p1", 1]]));
 
         expect(commit.mock.calls[4][0]).toBe(`model/${ModelMutation.SetPaletteModel}`);
@@ -428,7 +324,7 @@ describe("Model actions", () => {
         expect(commit.mock.calls[5][0]).toBe(`model/${ModelMutation.SetCompileRequired}`);
         expect(commit.mock.calls[5][1]).toBe(false);
 
-        expect(commit.mock.calls[6][0]).toBe(`model/${ModelMutation.SetRunRequired}`);
+        expect(commit.mock.calls[6][0]).toBe(`run/${RunMutation.SetRunRequired}`);
         expect(commit.mock.calls[6][1]).toBe(true);
 
         expect(commit.mock.calls[7][0]).toBe(`sensitivity/${SensitivityMutation.SetUpdateRequired}`);
@@ -444,9 +340,9 @@ describe("Model actions", () => {
         expect(run.mock.calls[0][2]).toBe(0); // start
         expect(run.mock.calls[0][3]).toBe(99); // end
 
-        expect(commit.mock.calls[9][0]).toBe(`model/${ModelMutation.SetOdinSolution}`);
+        expect(commit.mock.calls[9][0]).toBe(`run/${RunMutation.SetOdinSolution}`);
         expect(commit.mock.calls[9][1]).toBe("test solution");
-        expect(commit.mock.calls[10][0]).toBe(`model/${ModelMutation.SetRunRequired}`);
+        expect(commit.mock.calls[10][0]).toBe(`run/${RunMutation.SetRunRequired}`);
         expect(commit.mock.calls[10][1]).toBe(false);
     });
 });
